feat: 공개/비공개 설정 컴포넌트 분리
